Receiver

+++ src/net/liplum/utils/WorldUtil.java
@@ -4,7 +4,7 @@ import mindustry.Vars;
 import mindustry.world.Block;
 
 public class WorldUtil {
-    public static float toDrawXY(Block block, float blockXY) {
+    public static float toDrawXY(Block block, int blockXY) {
         return block.offset + blockXY * Vars.tilesize;
     }
 }
